+++ core/src/main/java/cucumber/runtime/StepDefinitionMatch.java
@@ -65,8 +65,7 @@ public class StepDefinitionMatch extends Match {
         if (step.getRows() != null) argumentCount++;
         if (parameterTypes != null) {
             if (parameterTypes.size() != argumentCount) {
-                List<Argument> arguments = createArgumentsForErrorMessage(step);
-                throw new CucumberException("Arity mismatch. Declared parameters: " + parameterTypes + ". Matched arguments: " + arguments);
+                throw arityMismatch(parameterTypes, step);
             }
         } else {
             // Some backends, like ClojureBackend, don't know the arity and therefore pass in null.
@@ -116,6 +115,21 @@ public class StepDefinitionMatch extends Match {
         return result;
     }
 
+    private CucumberException arityMismatch(List<ParameterType> parameterTypes, Step step) {
+        List<Argument> arguments = createArgumentsForErrorMessage(step);
+        return new CucumberException(String.format(
+                "Arity mismatch: Step Definition '%s' with pattern /%s/ is declared with %s parameters%s. However, the gherkin step matched %s arguments%s. \nStep: %s%s",
+                stepDefinition.getLocation(),
+                stepDefinition.getPattern(),
+                parameterTypes.size(),
+                parameterTypes.isEmpty() ? "" : " " + parameterTypes.toString(),
+                arguments.size(),
+                arguments.isEmpty() ? "" : " " + arguments.toString(),
+                step.getKeyword(),
+                step.getName()
+        ));
+    }
+
     private List<Argument> createArgumentsForErrorMessage(Step step) {
         List<Argument> arguments = new ArrayList<Argument>(getArguments());
         if (step.getDocString() != null) {
@@ -134,7 +148,7 @@ public class StepDefinitionMatch extends Match {
     }
 
     private Object tableArgument(Step step, int argIndex, XStream xStream, String dateFormat) {
-        DataTable table = new DataTable(step.getRows(), new TableConverter(xStream, dateFormat));
+        DataTable table = new DataTable(step.getRows(), new TableConverter(xStream));
 
         Type listType = getGenericListType(argIndex);
         if (listType != null) {