+++ groovy/src/test/resources/cucumber/runtime/groovy/interpreted_stepdefs.groovy
@@ -7,4 +7,3 @@ Given(~'^I have (\\d+) cukes in my belly') { int cukes ->
     haveCukes(cukes)
     lastAte('cukes')
 }
-