+++ java/src/test/java/cucumber/runtime/java/JavaBackendTest.java
@@ -1,6 +1,7 @@
 package cucumber.runtime.java;
 
 import cucumber.fallback.runtime.java.DefaultJavaObjectFactory;
+import cucumber.runtime.CucumberException;
 import cucumber.runtime.Glue;
 import cucumber.runtime.HookDefinition;
 import cucumber.runtime.StepDefinition;
@@ -19,12 +20,20 @@ import static java.util.Arrays.asList;
 import static org.junit.Assert.assertEquals;
 
 public class JavaBackendTest {
+    @Test(expected = CucumberException.class)
+    public void doesnt_like_path_like_glue() {
+        ObjectFactory factory = new DefaultJavaObjectFactory();
+        JavaBackend backend = new JavaBackend(factory);
+        GlueStub world = new GlueStub();
+        backend.loadGlue(world, asList("cucumber/runtime/java/test"));
+    }
+
     @Test
     public void finds_step_definitions_by_scanning_for_annotations() {
         ObjectFactory factory = new DefaultJavaObjectFactory();
         JavaBackend backend = new JavaBackend(factory);
         GlueStub world = new GlueStub();
-        backend.loadGlue(world, asList("cucumber/runtime/java/test"));
+        backend.loadGlue(world, asList("cucumber.runtime.java.test"));
         backend.buildWorld();
         assertEquals(Stepdefs.class, factory.getInstance(Stepdefs.class).getClass());
     }